Upgrade dependencies and related fixes
    $ npm upgrade

The `crypto` package is now built-in, so there is no need to list it in
the dependencies.  But calls to `pbkdf2Sync()` now need an extra
parameter.

Mongoose needs another connection parameter to deal with discrepancies
between Mongoose and the underlying MongoDB driver.

Added an explicit dependency on the `acorn` package to deal with a
security vulnerability.  It is brought in through the `pug` package.
When they upgrade **their** dependencies, I will be alble to remove this
explicit dependency that I don't need, myself.